new layout for metadata.
