+++ sharding-jdbc-core/src/main/java/io/shardingjdbc/core/jdbc/core/resultset/GeneratedKeysResultSetMetaData.java
@@ -25,7 +25,7 @@ import java.sql.ResultSetMetaData;
 import java.sql.Types;
 
 /**
- * ResultSet meta for generated keys.
+ * ResultSet metadata for generated keys.
  * 
  * @author gaohongtao
  */
