+++ sharding-jdbc-core/src/main/java/io/shardingjdbc/core/jdbc/metadata/entity/ActualTableInformation.java
@@ -1,5 +1,6 @@
-package io.shardingjdbc.core.jdbc.meta.entity;
+package io.shardingjdbc.core.jdbc.metadata.entity;
 
+import io.shardingjdbc.core.rule.DataNode;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 
@@ -12,9 +13,7 @@ import lombok.RequiredArgsConstructor;
 @Getter
 public class ActualTableInformation {
     
-    private final String dataSourceName;
-    
-    private final String tableName;
+    private final DataNode dataNode;
     
     private final TableMeta tableMeta;
 }
