+++ sharding-jdbc-core/src/main/java/io/shardingjdbc/core/jdbc/metadata/handler/TableMetaHandlerFactory.java
@@ -1,26 +1,24 @@
-package io.shardingjdbc.core.jdbc.meta.handler;
+package io.shardingjdbc.core.jdbc.metadata.handler;
 
 import io.shardingjdbc.core.constant.DatabaseType;
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
-
 import javax.sql.DataSource;
 import java.sql.SQLException;
 
 /**
- * The factory of table meta handler.
+ * The factory of table metadata handler.
  *
  * @author panjuan
  */
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
 public final class TableMetaHandlerFactory {
-    
     /**
-     * To generate table meta handler by data type.
+     * To generate table metadata handler by data type.
      *
      * @param dataSource data source.
      * @param actualTableName actual table name.
-     * @return abstract table meta handler.
+     * @return abstract table metadata handler.
      * @throws SQLException SQL exception.
      */
     public static AbstractTableMetaHandler newInstance(final DataSource dataSource, final String actualTableName) throws SQLException {
@@ -32,5 +30,4 @@ public final class TableMetaHandlerFactory {
                 throw new UnsupportedOperationException(String.format("Cannot support database [%s].", databaseType));
         }
     }
-    
 }
