+++ sharding-proxy/src/main/java/io/shardingjdbc/proxy/transport/mysql/packet/command/statement/prepare/ComStmtPrepareOKPacket.java
@@ -53,7 +53,7 @@ public final class ComStmtPrepareOKPacket extends MySQLPacket {
         mysqlPacketPayload.writeInt1(status);
         mysqlPacketPayload.writeInt4(statementId);
         // TODO Set numColumns=0 is a workaround to escape jdbc check for now, there's no issues found during a few tests.
-        // TODO Column Definition Block should be added in future when the meta of the columns is cached.
+        // TODO Column Definition Block should be added in future when the metadata of the columns is cached.
         mysqlPacketPayload.writeInt2(0);
         mysqlPacketPayload.writeInt2(numParameters);
         mysqlPacketPayload.writeReserved(1);
